little fix

+++ src/main/java/com/Payment/Shop/dto/request/VNPaymentRequest.java
@@ -1,10 +1,12 @@
 package com.Payment.Shop.dto.request;
 
+import java.math.BigDecimal;
+
 import lombok.*;
 
 @Getter
 @Setter
 @NoArgsConstructor
 public class VNPaymentRequest extends PaymentRequest{
-    protected double totalAmount;
+    protected BigDecimal totalAmount;
 }
